# Optimize code and string allocations