--- a/jsbridge/build.gradle
+++ b/jsbridge/build.gradle
@@ -12,6 +12,7 @@ group = 'de.prosiebensat1digital'
 
 android {
     compileSdkVersion rootProject.ext.versions.compile_sdk
+    namespace "de.prosiebensat1digital.oasisjsbridge"
 
     defaultConfig {
         targetSdkVersion rootProject.ext.versions.target_sdk
@@ -50,6 +51,10 @@ android {
             test.java.srcDirs += ['src/test/kotlin']
             androidTest.java.srcDirs += ['src/androidTest/kotlin']
         }
+    }
+
+    buildFeatures {
+        buildConfig true
     }
 
     externalNativeBuild {