--- a/jsbridge/src/main/jni/java-types/Object.cpp
+++ b/jsbridge/src/main/jni/java-types/Object.cpp
@@ -270,6 +270,7 @@ JavaType *Object::newJavaType(const JniLocalRef<jobject> &object) const {
     default:
       return nullptr;
   }
+  return nullptr;
 }
 
 }  // namespace JavaTypes