… jsbridge/src/main/assets/js/promises.js
… jsbridge/src/main/kotlin/de/prosiebensat1digital/oasisjsbridge/PayloadObject.kt
… jsbridge/src/main/kotlin/de/prosiebensat1digital/oasisjsbridge/extensions/PromiseExtension.kt
… jsbridge/src/main/kotlin/de/prosiebensat1digital/oasisjsbridge/extensions/XMLHttpRequestExtension.kt
… jsbridge/src/test/kotlin/de/prosiebensat1digital/oasisjsbridge/JsonObjectWrapperTest.kt
… jsbridge/src/test/kotlin/de/prosiebensat1digital/oasisjsbridge/PayloadObjectTest.kt