- Used `StringBuilder` to construct json strings instead of keeping N copies of the entire string in memory. This makes `JsonObjectWrapper` and `PayloadObject` a lot faster and use less memory
- Load JS code from assets instead of using static strings in memory


Benchmark test on my Pixel 6:

`JsonObjectWrapper.collectionToJsonString`, 1000 elements:

```
Before: 2964 ms, After: 106 ms
```

`JsonObjectWrapper.mapToJsonString`, 1000 properties:

```
Before: 1014 ms, After: 55 ms
```