@bwalter089 2024-02-29
The re-format (which makes sense) and the move to truth (I am not sure about the benefit here ;)) make it a little bit hard to review the real changes but the optimizations are very welcome additions, thanks a lot for your work! Looks good to me,